requirements/base.txt
# Core
pytz==2019.2
Django==2.2.8 # pyup: < 3.0
Pillow==6.2.0

# Configuration manager 
django-configurations==2.1
python-decouple==3.1

# For the persistence stores
psycopg2-binary==2.8.3
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.2.0

# Rest apis
djangorestframework==3.10.3
Markdown==3.1.1
django-filter==2.2.0
django-cors-headers==3.1.1
django-extra-fields==1.2.4
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
